+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 ### 🚀 Features
 
 - ## **playground**:
-  - Custom border type (solid, dashed, dotted, double), color, width, and opacity - by @thuongtruong109 [<samp>()</samp>]()
+  - Custom editor border style(type, color, width, and opacity) - by @thuongtruong109 [<samp>()</samp>]()
 
 ##### [View changes on GitHub](https://github.com/thuongtruong109/flashot/compare/v2.4.0...v2.5.0) - [Full Changelog](https://github.com/thuongtruong109/flashot/commits/v2.5.0)